refactoring: Moved interfaces
