+++ src/DataTransferObjects/GenderStatisticsDto.php
@@ -2,6 +2,9 @@
 
 namespace App\DataTransferObjects;
 
+use App\Application\Contract\ConstructFromArrayInterface;
+use App\Application\Traits\ConstructableFromArrayTrait;
+
 class GenderStatisticsDto implements ConstructFromArrayInterface
 {
     use ConstructableFromArrayTrait;
